Do not rewrite Google Sheets links during URL formatting
The URL formatting logic that converts Google Docs links to
Library links was a little too eager as it rewrote Google Sheets
links as well. As a result, the fragment part of the URLs was
lost. This was annoying, because the fragment is often used to
point the user to edit a particular tab of the sheet.

Now the formatting simply ignores Google Sheets links and leaves
them as they are.